split up integration tests into gateway and auth
