+++ contracts/axelar-gateway/src/testutils.rs
@@ -53,6 +53,17 @@ pub fn setup_gateway<'a>(
     (signer_set, client)
 }
 
+pub fn setup_env<'a>(
+    previous_signers_retention: u32,
+    num_signers: u32,
+) -> (Env, TestSignerSet, AxelarGatewayClient<'a>) {
+    let env = Env::default();
+    env.mock_all_auths();
+    let (signers, client) = setup_gateway(&env, previous_signers_retention, num_signers);
+
+    (env, signers, client)
+}
+
 pub fn get_approve_hash(env: &Env, messages: Vec<Message>) -> BytesN<32> {
     env.crypto()
         .keccak256(&(CommandType::ApproveMessages, messages).to_xdr(env))
